frontend: ArticleTable: refactor article control popup.
